+++ Q3E/src/main/jni/doom3/neo/sys/android/main.cpp
@@ -695,7 +695,7 @@ static void Q3E_FreeArgs(void)
 	free(q3e_argv);
 }
 
-// doom3 game main thread loop
+// idTech4 game main thread loop
 static void * doom3_main(void *data)
 {
 	attach_thread(); // attach current to JNI for call Android code
@@ -708,7 +708,7 @@ static void * doom3_main(void *data)
 	Q3E_Start();
 
 	Posix_EarlyInit();
-	Sys_Printf("[Harmattan]: Enter doom3 main thread -> %s\n", Sys_GetThreadName());
+	Sys_Printf("[Harmattan]: Enter idTech4 main thread -> %s\n", Sys_GetThreadName());
 
 	if (q3e_argc > 1) {
 		common->Init(q3e_argc-1, (const char **)&q3e_argv[1], NULL);
@@ -729,7 +729,7 @@ static void * doom3_main(void *data)
 	common->Quit();
 	Q3E_End();
 	main_thread = 0;
-	Sys_Printf("[Harmattan]: Leave doom3 main thread.\n");
+	Sys_Printf("[Harmattan]: Leave idTech4 main thread.\n");
 	return 0;
 }
 
@@ -743,19 +743,19 @@ static void Q3E_StartGameMainThread(void)
 	pthread_attr_init(&attr);
 
 	if (pthread_attr_setdetachstate(&attr, PTHREAD_CREATE_JOINABLE) != 0) {
-		Sys_Printf("[Harmattan]: ERROR: pthread_attr_setdetachstate doom3 main thread failed\n");
+		Sys_Printf("[Harmattan]: ERROR: pthread_attr_setdetachstate idTech4 main thread failed\n");
 		exit(1);
 	}
 
 	if (pthread_create((pthread_t *)&main_thread, &attr, doom3_main, NULL) != 0) {
-		Sys_Printf("[Harmattan]: ERROR: pthread_create doom3 main thread failed\n");
+		Sys_Printf("[Harmattan]: ERROR: pthread_create idTech4 main thread failed\n");
 		exit(1);
 	}
 
 	pthread_attr_destroy(&attr);
 
 	q3e_running = true;
-	Sys_Printf("[Harmattan]: doom3 main thread start.\n");
+	Sys_Printf("[Harmattan]: idTech4 main thread start.\n");
 }
 
 // shutdown game main thread
@@ -766,17 +766,17 @@ static void Q3E_ShutdownGameMainThread(void)
 
 	q3e_running = false;
 	if (pthread_join(main_thread, NULL) != 0) {
-		Sys_Printf("[Harmattan]: ERROR: pthread_join doom3 main thread failed\n");
+		Sys_Printf("[Harmattan]: ERROR: pthread_join idTech4 main thread failed\n");
 	}
 	main_thread = 0;
-	Sys_Printf("[Harmattan]: doom3 main thread quit.\n");
+	Sys_Printf("[Harmattan]: idTech4 main thread quit.\n");
 }
 
 void ShutdownGame(void)
 {
     if(common->IsInitialized())
     {
-        Sys_TriggerEvent(TRIGGER_EVENT_WINDOW_CREATED); // if doom3 main thread is waiting new window
+        Sys_TriggerEvent(TRIGGER_EVENT_WINDOW_CREATED); // if idTech4 main thread is waiting new window
         Q3E_ShutdownGameMainThread();
         common->Quit();
     }
@@ -784,7 +784,7 @@ void ShutdownGame(void)
 
 static void doom3_exit(void)
 {
-	Sys_Printf("[Harmattan]: doom3 exit.\n");
+	printf("[Harmattan]: idTech4 exit.\n");
 
 	Q3E_CloseRedirectOutput();
 }